linux: remove IN_ONESHOT and handle inotify watches cleanly.
Previously the inotify event watch for ever file was removed immediately
after an event has happened, so inotify_add_watch needed to be called
every time. This was redundant and it's sufficient to call it only when
the inode of the file has changed (eg. in IN_MOVE_SELF event). In all
other cases - keep the original watch in place.