Add a teardown command option.
In case the command execution reaches the timeout, instead of killing
the process, a teardown command will be run if specified. This is useful
if the command produces multiple subprocesses in which case killing of
the child process may produce zombie processes. So it would be more
efficient to run something like 'pkill <program>.
A $CMD_PID environment variable is populated for the teardown command,
with the pid of the command subprocess.